+++ cypress/deprecated/structured-search.spec.js
@@ -9,7 +9,7 @@ import { ROUTE_STRUCTURED_SEARCH } from '../../src/constants/routes'
 
 context('Structured search', () => {
   beforeEach(() => {
-    cy.visit('http://localhost:3000')
+    cy.visit('/')
   })
 
   describe('Structured search', () => {